feat: add primary logic and measure times

+++ src/main/java/pp/finki/ukim/mk/annocuda/kernels/vectorAdd.cu
@@ -1,6 +1,6 @@
 extern "C"
 __global__
-void add (int n, float *a, float *b){
+void add (long n, double *a, double *b){
     int i = blockIdx.x * blockDim.x + threadIdx.x;
     if (i < n)
     {
